755: Fix endpoint authorization matls checks
Matls used to work like this;
spring-security-multi-auth would look up the tpp using the certificateCN
(only worked for OB Transport certs), and put the oidc client from teh
Tpp record into the spring principal name. Then the Endpoint Wrapper
would look at the access_token and get the oauth2 clientId from there
and check it was the same as the spring principal name.

However, the presented cert can now not be tied to a specific OAuth2
client Id. All the spring-security-multi-auth can do is populate the
principal name with the authorizationNumber taken from the PSD2 eIDAS
certificate. So, the endpoint wrapper must look for tpp based on the
OAuth2 Client Id (unique to a Tpp record) and check that the
authorizationNumber in that TPP record matches the Spring principal
name. This will ensure that the access_token was issued to an OAuth2
client that belongs to the TPP as identified in the presented transport
certificate.

Issue: https://github.com/ForgeCloud/ob-deploy/issues/775