830: Manual Registation now works
Addressing the issue here that the register app had to call the as-api
manual registration endpoint, which in turn created the TPP.

Better to do all the checks in the registration application and create
the TPP and ManualRegistration application from there.

We expect the browser to be configured to use an OBWac belonging to the
same OB Directory organisation as found in the SSA used in the
registration.

The ManualRegistrationApplication store in mongo is used to allow us to
record which logged in user owns which applications. A logged in user
can only see their own applications.

Issue: https://github.com/ForgeCloud/ob-deploy/issues/830