859: Fix import of CustomerInfo data
When created via the import apis `/api/user/data`, the CustomerInfo did
not have the required `username` field in it. This needs to be added
from the UserData recived with the data. It's absence meant that the RCS
could not find the CustomerInfo for a specific PSU.

Issue: https://github.com/ForgeCloud/ob-deploy/issues/859